add missing param to openapi

+++ packages/site/src/api/openapi/apis/RolesApi.ts
@@ -34,6 +34,7 @@ export interface RolesApiCreateTenantRoleRequest {
 
 export interface RolesApiRemoveRequest {
 	id: string;
+	organizationId: string;
 }
 
 export interface RolesApiSendInviteRequest {
@@ -396,16 +397,31 @@ export class RolesApi extends runtime.BaseAPI implements RolesApiInterface {
 			);
 		}
 
+		if (
+			requestParameters.organizationId === null ||
+			requestParameters.organizationId === undefined
+		) {
+			throw new runtime.RequiredError(
+				'organizationId',
+				'Required parameter requestParameters.organizationId was null or undefined when calling remove.',
+			);
+		}
+
 		const queryParameters: any = {};
 
 		const headerParameters: runtime.HTTPHeaders = {};
 
 		const response = await this.request(
 			{
-				path: `/organizations/{organizationId}/roles/{id}`.replace(
-					`{${'id'}}`,
-					encodeURIComponent(String(requestParameters.id)),
-				),
+				path: `/organizations/{organizationId}/roles/{id}`
+					.replace(
+						`{${'id'}}`,
+						encodeURIComponent(String(requestParameters.id)),
+					)
+					.replace(
+						`{${'organizationId'}}`,
+						encodeURIComponent(String(requestParameters.organizationId)),
+					),
 				method: 'DELETE',
 				headers: headerParameters,
 				query: queryParameters,
